+++ react/Admin.tsx
@@ -10,11 +10,8 @@ import {
   ButtonPlain,
   Spinner,
   Divider,
-  Tooltip,
-  IconHelp,
 } from 'vtex.styleguide'
-import type { WrappedComponentProps } from 'react-intl'
-import { injectIntl, FormattedMessage } from 'react-intl'
+import { FormattedMessage, useIntl } from 'react-intl'
 import { useQuery, useMutation } from 'react-apollo'
 
 import GoogleSignIn from '../public/metadata/google_signin.png'
@@ -23,14 +20,15 @@ import Q_HAVE_TOKEN from './queries/HaveToken.gql'
 import Q_SHEET_LINK from './queries/SheetLink.gql'
 import M_REVOKE from './mutations/RevokeToken.gql'
 import M_CREATE_SHEET from './mutations/CreateSheet.gql'
-import M_PROCESS_SHEET from './mutations/ProcessSheet.gql'
-import M_CLEAR_SHEET from './mutations/ClearSheet.gql'
-import M_ADD_IMAGES from './mutations/AddImages.gql'
+import ProcessSheetButton from './components/ProcessSheetButton'
+import ClearSheetButton from './components/ClearSheetButton'
+import AddImagesButton from './components/AddImagesButton'
 
 const AUTH_URL = '/sheets-catalog-import/auth'
 
-const Admin: FC<WrappedComponentProps> = ({ intl }) => {
+const Admin: FC = () => {
   const { account, pages } = useRuntime()
+  const intl = useIntl()
 
   const {
     loading: ownerLoading,
@@ -67,9 +65,6 @@ const Admin: FC<WrappedComponentProps> = ({ intl }) => {
     { loading: createLoading, data: createData, called: createCalled },
   ] = useMutation(M_CREATE_SHEET)
 
-  const [sheetImport, { loading: sheetProcessing, data: sheetProcessed }] =
-    useMutation(M_PROCESS_SHEET)
-
   const auth = () => {
     revoke()
       .then(() => {
@@ -86,12 +81,6 @@ const Admin: FC<WrappedComponentProps> = ({ intl }) => {
       (createCalled && !createLoading && !!createData?.createSheet)
     )
   }
-
-  const [clearSheet, { loading: sheetClearing, data: sheetCleared }] =
-    useMutation(M_CLEAR_SHEET)
-
-  const [addImages, { loading: addingImages, data: imagesAdded }] =
-    useMutation(M_ADD_IMAGES)
 
   return (
     <Layout
@@ -225,142 +214,12 @@ const Admin: FC<WrappedComponentProps> = ({ intl }) => {
             </Card>
           )}
           <br />
-          {showLink() && (
-            <div>
-              <Card>
-                <div className="flex">
-                  <div className="w-70">
-                    <p>
-                      <FormattedMessage id="admin/sheets-catalog-import.sheet-import.description" />
-                    </p>
-                  </div>
-                  <div
-                    style={{ flexGrow: 1 }}
-                    className="flex items-stretch w-20 justify-center"
-                  >
-                    <Divider orientation="vertical" />
-                  </div>
-                  <div className="w-30 items-center flex">
-                    {!sheetProcessed?.processSheet && (
-                      <Button
-                        variation="secondary"
-                        collapseLeft
-                        block
-                        isLoading={sheetProcessing}
-                        onClick={() => {
-                          sheetImport()
-                        }}
-                      >
-                        <FormattedMessage id="admin/sheets-catalog-import.sheet-import.button" />
-                      </Button>
-                    )}
-                    {!sheetProcessing && sheetProcessed?.processSheet && (
-                      <p>
-                        <strong>{`${sheetProcessed.processSheet}`}</strong>
-                      </p>
-                    )}
-                  </div>
-                </div>
-              </Card>
-              <br />
-            </div>
-          )}
+          {showLink() && <ProcessSheetButton />}
           <br />
-          {showLink() && (
-            <div>
-              <Card>
-                <div className="flex">
-                  <div className="w-70">
-                    <p>
-                      <FormattedMessage id="admin/sheets-catalog-import.sheet-clear.description" />
-                    </p>
-                  </div>
-                  <div
-                    style={{ flexGrow: 1 }}
-                    className="flex items-stretch w-20 justify-center"
-                  >
-                    <Divider orientation="vertical" />
-                  </div>
-                  <div className="w-30 items-center flex">
-                    {!sheetCleared?.clearSheet && (
-                      <Button
-                        variation="secondary"
-                        collapseLeft
-                        block
-                        isLoading={sheetClearing}
-                        onClick={() => {
-                          clearSheet()
-                        }}
-                      >
-                        <FormattedMessage id="admin/sheets-catalog-import.sheet-clear.button" />
-                      </Button>
-                    )}
-                    {!sheetClearing && sheetCleared?.clearSheet && (
-                      <p>
-                        <strong>{`${sheetCleared.clearSheet}`}</strong>
-                      </p>
-                    )}
-                  </div>
-                </div>
-              </Card>
-              <br />
-            </div>
-          )}
+          {showLink() && <ClearSheetButton />}
           <br />
           {pages['admin.app.google-drive-import'] && showLink() && (
-            <div>
-              <Card>
-                <div className="flex">
-                  <div className="w-70">
-                    <p>
-                      <div>
-                        <span className="mr5">
-                          <FormattedMessage id="admin/sheets-catalog-import.add-images.description" />
-                        </span>
-                        <Tooltip
-                          label={intl.formatMessage({
-                            id: 'admin/sheets-catalog-import.add-images.tooltip',
-                            defaultMessage:
-                              'This feature only works if you are also using the VTEX Google Drive App, which creates a folder in your drive. You can use the folder titled NEW to add product images. These images will be prepopulated in your Google Catalog Import spreadsheet.',
-                          })}
-                        >
-                          <span className="ml-4 c-on-base pointer">
-                            <IconHelp />
-                          </span>
-                        </Tooltip>
-                      </div>
-                    </p>
-                  </div>
-                  <div
-                    style={{ flexGrow: 1 }}
-                    className="flex items-stretch w-20 justify-center"
-                  >
-                    <Divider orientation="vertical" />
-                  </div>
-                  <div className="w-30 items-center flex">
-                    {!imagesAdded?.addImages && (
-                      <Button
-                        variation="secondary"
-                        collapseLeft
-                        block
-                        isLoading={addingImages}
-                        onClick={() => {
-                          addImages()
-                        }}
-                      >
-                        <FormattedMessage id="admin/sheets-catalog-import.add-images.button" />
-                      </Button>
-                    )}
-                    {!addingImages && imagesAdded?.addImages && (
-                      <p>
-                        <strong>{`${imagesAdded.addImages}`}</strong>
-                      </p>
-                    )}
-                  </div>
-                </div>
-              </Card>
-              <br />
-            </div>
+            <AddImagesButton />
           )}
         </div>
       )}
@@ -368,4 +227,4 @@ const Admin: FC<WrappedComponentProps> = ({ intl }) => {
   )
 }
 
-export default injectIntl(Admin)
+export default Admin